requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-redis==5.4.0

uvicorn==0.32.0

channels==4.1.0
channels-redis==4.2.0

whitenoise==6.8.2

psycopg2== 2.9.4 #db

django-environ==0.9.0 #variables de ambiente

pillow==11.0.0 #para procesar imagenes

django-ckeditor-5==0.2.15 #editor para el content

celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.7.0

djangorestframework-api-response==0.1.0 #Para una respuesta mas estandarizada (por ejemplo con el status, etc, parecido a lo que hacia con schema generico) creada por el prof de udemy